OCR_Automator/requirements_web.txt
Flask>=3.0.0,<4
pytesseract>=0.3.10
pdf2image>=1.17.0
pillow>=10.0.0
pandas>=2.2.0
openpyxl>=3.1.2
requests>=2.31.0
